Move `__version__` to `__init__.py`
