+++ e3sm_to_cmip/__init__.py
@@ -1,0 +1,3 @@
+"""Top-level package for e3sm_to_cmip."""
+
+__version__ = "1.8.2"
