--- e3sm_to_cmip/version.py
@@ -1,1 +1,0 @@
-__version__ = "1.8.2"
