+++ tbump.toml
@@ -21,7 +21,7 @@ tag_template = "v{new_version}"
 # section containing the path of the file, relative to the
 # tbump.toml location.
 [[file]]
-src = "e3sm_to_cmip/version.py"
+src = "e3sm_to_cmip/__init__.py"
 
 # You can specify a list of commands to
 # run after the files have been patched
